Moved some things where they belong
