+++ src/main/java/tk/estecka/shiftingwares/TradeLayouts/VanillaTradeLayout.java
@@ -5,6 +5,7 @@ import java.util.List;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.minecraft.entity.passive.VillagerEntity;
 import net.minecraft.resource.featuretoggle.FeatureFlags;
+import net.minecraft.util.Identifier;
 import net.minecraft.village.TradeOffers;
 import net.minecraft.village.VillagerData;
 import net.minecraft.village.VillagerProfession;
@@ -15,13 +16,22 @@ import tk.estecka.shiftingwares.api.ITradeLayoutProvider;
 public class VanillaTradeLayout
 implements ITradeLayoutProvider
 {
+	static public final Identifier TRADE_REBALANCE_FLAGID = new Identifier("minecraft", "trade_rebalance");
+	static public final boolean IS_EXP_TRADE_AVAILABLE;
+
+	static {
+		var featureSet = FeatureFlags.FEATURE_MANAGER.getFeatureSet();
+		var featureIds = FeatureFlags.FEATURE_MANAGER.toId(featureSet);
+		IS_EXP_TRADE_AVAILABLE = featureIds.contains(TRADE_REBALANCE_FLAGID);
+	}
+
 	public List<Factory[]>	GetTradeLayout(VillagerEntity villager){
 		List<Factory[]> layout = new ArrayList<>();
 		VillagerProfession job = villager.getVillagerData().getProfession();
 		int jobLevel = villager.getVillagerData().getLevel();
 
 		Int2ObjectMap<Factory[]> jobPool = TradeOffers.PROFESSION_TO_LEVELED_TRADE.get(job);
-		if (ShiftingWares.IS_EXP_TRADE_AVAILABLE 
+		if (IS_EXP_TRADE_AVAILABLE 
 		&& villager.getWorld().getEnabledFeatures().contains(FeatureFlags.TRADE_REBALANCE)
 		&& TradeOffers.REBALANCED_PROFESSION_TO_LEVELED_TRADE.containsKey(job)
 		) {
